(#441) code 动态2.0发布动态后台初步调整

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/repository/BaseDynamicRepository.java
@@ -660,7 +660,7 @@ public class BaseDynamicRepository implements IDynamicReppsitory {
                                                 userInfoBeanSparseArray.put(userInfoBean.getUser_id().intValue(), userInfoBean);
                                             }
                                             for (DynamicDetailBeanV2 dynamicBean : listBaseJson) {
-                                                dynamicBean.setUserInfoBean(userInfoBeanSparseArray.get(dynamicBean.getUser_id()));
+                                                dynamicBean.setUserInfoBean(userInfoBeanSparseArray.get(dynamicBean.getUser_id().intValue()));
                                                 for (int i = 0; i < dynamicBean.getComments().size(); i++) {
                                                     dynamicBean.getComments().get(i).setCommentUser(userInfoBeanSparseArray.get((int) dynamicBean.getComments().get(i).getUser_id()));
                                                     if (dynamicBean.getComments().get(i).getReply_to_user_id() == 0) { // 如果 reply_user_id = 0 回复动态
